refactor(backend): 의존성 경계와 API 조립 구조 정리
