+++ src/google_work_agent/api/__init__.py
@@ -1,5 +1,6 @@
 """FastAPI surface for the local product core."""
 
-from google_work_agent.api.app import API_CONTRACT_VERSION, ApiContainer, create_app
+from google_work_agent.api.app import create_app
+from google_work_agent.api.container import API_CONTRACT_VERSION, ApiContainer
 
 __all__ = ["API_CONTRACT_VERSION", "ApiContainer", "create_app"]
