+++ src/google_work_agent/application/workflows/request_understanding.py
@@ -6,9 +6,9 @@ from functools import partial
 from pathlib import Path
 from typing import Final, Literal, NotRequired, Required, TypedDict, cast
 
+import google_work_agent.application.workflows._schema_support as _schema
 from google_work_agent.application.llm import StructuredLLMRuntime
 from google_work_agent.application.observability import ObservabilityContext
-from google_work_agent.application.workflows import _schema_support as _schema
 from google_work_agent.application.workflows.contracts import (
     CONFIRMATION_ORIGIN_TARGETS,
     ConfirmationResponseV1,
